esp_h264: Fix slice header wrongly set

+++ esp_h264/hw/src/esp_h264_enc_hw_param.h
@@ -36,7 +36,7 @@ typedef struct esp_h264_set_cfg {
     uint8_t            qp_min;   /*<! The minimum quantization parameter(QP) */
     uint8_t            qp_max;   /*<! The maximum quantization parameter(QP) */
     uint8_t            fps;      /*<! Frames per second */
-    uint8_t            bitrate;  /*<! Bit per second */
+    uint32_t           bitrate;  /*<! Bit per second */
 } esp_h264_enc_hw_param_cfg_t;
 
 /**
